requirements.txt

gradio==3.42.0
tensorflow==2.15.0
numpy==1.25.0
Pillow==10.0.0
gTTS==2.3.2
googletrans==4.0.0-rc1
requests==2.31.0
